Lift main-loop generator
